+++ src/Sulu/Bundle/SecurityBundle/Resources/config/services.xml
@@ -281,11 +281,16 @@
         </service>
 
         <service id="sulu_security.twig_extension.user.cache" class="Doctrine\Common\Cache\ArrayCache"/>
+
         <service id="sulu_security.twig_extension.user" class="Sulu\Bundle\SecurityBundle\Twig\UserTwigExtension">
+            <tag name="twig.extension"/>
+        </service>
+
+        <service id="Sulu\Bundle\SecurityBundle\Twig\UserRuntime">
             <argument type="service" id="sulu_security.twig_extension.user.cache"/>
             <argument type="service" id="sulu.repository.user"/>
 
-            <tag name="twig.extension"/>
+            <tag name="twig.runtime"/>
         </service>
 
         <service id="sulu_security.user_locale_listener"
